pyproject.toml
[project]
name = "rns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langdetect==1.0.9",
    "musicbrainzngs==0.7.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
